[refactor] Combine LectureSWQueryRepo and LectureSWRepo
